requirements.txt
numpy==1.24.4
Pillow==10.3.0
zipp==3.19.2
Flask==3.0.2
Flask-Cors==4.0.1
Jinja2==3.1.2
Werkzeug==3.0.1
MarkupSafe>=2.1.2
opencv-python==4.10.0.84
openslide-python==1.3.1
flaskwebgui==1.1.2
appdirs==1.4.4